пр 14 проблема с селебрейт

+++ controllers/cards.js
@@ -22,7 +22,7 @@ module.exports.postCards = (req, res) => {
 };
 
 module.exports.deleteCards = (req, res) => {
-  Card.deleteOne({ _id: req.params.cardId })
+  Card.deleteOne({ _id: req.params.cardId, owner: req.user._id })
     .then((deleteStatus) => {
       if (deleteStatus.deletedCount === 0) {
         return res.status(404).send({ message: 'Карточка с указанным _id не найдена.' });
